Move controller to lib
