--- packages/viewer/src/scene-tree/index.ts
@@ -1,1 +1,0 @@
-export * from './row';
